requirements.txt
robotframework==7.4.1
robotframework-appiumlibrary==3.2.1
robotframework-faker==6.0.0
